🧹 Tidying up responses
